refactor(settings): split LayoutDrift out of SettingsModel
The #403 merge (seedSpaces + its doc comments) pushed
SettingsModel.swift to 355 lines, past the 350 ceiling — its
own gate ran before the two additions met on main. The drift
snapshot type and its compute/refresh logic move to
SettingsModel+LayoutDrift.swift; the published property stays
with the model (a stored property cannot live in an extension).
